used the helper functions to clean code, added some even more extra security measures and error handlings

+++ controllers/profile/DeleteUserController.js
@@ -1,6 +1,7 @@
 
 
-const {deleteUser} = require("../../database/queries");
+const {deleteUser, userExistsById} = require("../../database/queries");
+const isNumeric = require("../../helper/isNumeric");
 
 
 
@@ -11,22 +12,41 @@ const DeleteUserController = async function(req, res){
     const {userinfo} = req.params;
     const userid = userinfo.split(",")[0];
     const fromUser = userinfo.split(",")[1];
-    if (fromUser == "true"){
-        await deleteUser(userid);
-        res.redirect("/homepage");
+    if (isNumeric(userid)){
+        if (fromUser == "true"){
+            if (req.user.isadmin){
+                const result = await deleteUser(userid);
+                if (result)
+                   res.redirect("/homepage");
+                else
+                    res.status(404).render("errors", {errors: [{msg : "User doesn't exist"}]});
+            }
+            else{
+                return res.status(500).render("errors", {errors: [{msg : "Deletion unsuccessful, You are not an Admin!"}]});
+            }
+    }
+        else{
+            const exists = await userExistsById(userid);
+            if (exists){
+                req.logout(err => {
+                    if (err) return res.status(500).render("errors", {errors : [err]});
+
+                    req.session.destroy(async err => {
+                        if (err) return res.status(500).render("errors", {errors : [err]});
+
+                        await deleteUser(userid);
+                        res.clearCookie('connect.sid');
+                        res.redirect('/login');
+                    });
+                    });
+            }
+            else{
+                res.status(500).render("errors", {errors: [{msg : "Deletion unsuccessful, You are not an Admin!"}]});
+            }
+        }
     }
     else{
-         req.logout(err => {
-        if (err) return res.status(500).render("errors", {errors : [err]});
-
-        req.session.destroy(async err => {
-            if (err) return res.status(500).render("errors", {errors : [err]});
-
-            await deleteUser(userid);
-            res.clearCookie('connect.sid');
-            res.redirect('/login');
-        });
-        });
+        res.status(404).render("errors", {errors : [{msg : "Id is not a number."}]})
     }
 }
